stock/easystock/requirements.txt
pytube 
youtube_transcript_api 
google-api-python-client
python-dotenv
openai-whisper
snownlp
